poc/requirements.txt
autogen-agentchat
autogen-core
azure-ai
azure-search-documents
asyncio
